requirements.txt
# Core LLM / Agent frameworks
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
openai>=1.30.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Web search
tavily-python>=0.3.0

# PDF parsing
pypdf>=4.0.0

# Config & environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Data & evaluation
pandas>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
tqdm>=4.66.0

# Visualizations
matplotlib>=3.8.0
seaborn>=0.13.0

# Notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
